Terminado el output del primer ejercicio

+++ src/include/misc.hpp
@@ -4,7 +4,7 @@
 #include "grafo.hpp"
 
 void imprimir_agm(listAristas l);
-listAristas remover_inconsistentes(listAristas l, Grafo g, int ds, double f, double diametro, int mod);
+listAristas remover_inconsistentes(listAristas l, Grafo g, double ds, double f, double diametro, int mod);
 void descubrirConexoAux(int u, listAristas res, int& contador, std::vector<bool> &visitado, int& i, std::vector<int> &comp_conex);
 std::vector<int> descubrirConexo(listAristas l, int n);
 listAristas diff(listAristas& l1, listAristas& l2);
